requirements.txt
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
Authlib==1.6.6
bandit==1.9.2
boolean.py==5.0
CacheControl==0.14.4
cachetools==6.2.3
certifi==2025.11.12
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
cyclonedx-python-lib==11.6.0
defusedxml==0.7.1
detect-secrets==1.5.0
disposable-email-domains==0.0.153
distlib==0.4.0
dj-database-url==3.0.1
Django==6.0
django-ckeditor==6.7.3
django-csp==4.0
django-htmx==1.27.0
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-mathfilters==1.0.0
django-ratelimit==4.1.0
django-recaptcha==4.1.0
django-storages==1.14.6
dparse==0.6.4
filelock==3.20.1
gevent==25.9.1
git-filter-repo==2.47.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-cloud-core==2.5.0
google-cloud-storage==3.7.0
google-crc32c==1.7.1
google-generativeai==0.8.5
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
identify==2.6.15
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
Levenshtein==0.27.3
license-expression==30.4.4
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.1.1
mdurl==0.1.2
msgpack==1.1.2
nltk==3.9.2
nodeenv==1.10.0
packageurl-python==0.17.6
packaging==25.0
pillow==12.0.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.10.0
platformdirs==4.5.1
pre_commit==4.5.1
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.11
py-serializable==2.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
python-dotenv==1.2.1
python-Levenshtein==0.27.3
PyYAML==6.0.3
qrcode==8.2
RapidFuzz==3.14.3
regex==2025.11.3
reportlab==4.4.7
requests==2.32.5
rich==14.2.0
rsa==4.9.1
ruamel.yaml==0.18.17
ruamel.yaml.clib==0.2.15
safety==3.7.0
safety-schemas==0.0.16
shellingham==1.5.4
sortedcontainers==2.4.0
sqlparse==0.5.5
stevedore==5.6.0
tenacity==9.1.2
tomli==2.3.0
tomli_w==1.2.0
tomlkit==0.13.3
tqdm==4.67.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.1
virtualenv==20.35.4
whitenoise==6.11.0
zope.event==6.1
zope.interface==8.1.1
